Bouncers: fix find_progressions
find_progressions would terminate a walk through the current `i mod k`
prematurely if there were less than 3 records with the same state
and direction.

Fixing this doesn't affect which machines are decided on the currently
undecided subset of the database, but when only using the deciders in
this repository, it decides 8 machines more.

As a result of this fix, the decider is about 2 times slower.